backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
